+++ src/snippets/lua/Action Wheel/Left Click Action.lua
@@ -1,0 +1,10 @@
+function pings.${2:action}Clicked()
+	$0
+end
+
+local ${2:action} = ${1:page}:newAction()
+	:title("$3")
+	:item("minecraft:$4")
+	:onLeftClick(function()
+		pings.${2:action}Clicked()
+	end)