+++ src/snippets/lua/General/Json Text Component.lua
@@ -1,0 +1,2 @@
+--? Json text component for within a json string. Use `Json Text` snippet to also generate the actual string around it.
+{"text"="$1", color="$2"}$0